# Conflicts:
#	Parameters.json
#	teamup-input/src/main/java/ru/team/up/input/controller/privateController/ApplicationController.java
#	teamup-sup/src/main/java/ru/team/up/sup/service/ParameterServiceImp.java

+++ teamup-input/src/main/java/ru/team/up/input/controller/privateController/UserController.java
@@ -26,6 +26,7 @@ import ru.team.up.sup.service.ParameterService;
 import javax.persistence.PersistenceException;
 import javax.validation.constraints.NotNull;
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -96,7 +97,7 @@ public class UserController {
                 UserDtoResponse.builder().
                         userDto(UserMapper.INSTANCE.mapUserToDto(user)).build(),
                 HttpStatus.OK);
-        Map<String, Object> monitoringParameters = new HashMap<>();
+        Map<String, Object> monitoringParameters = new LinkedHashMap<>();
         monitoringParameters.put("ID ", user.getId());
         monitoringParameters.put("Email ", user.getEmail());
         monitoringParameters.put("Имя ", user.getUsername());
@@ -130,6 +131,16 @@ public class UserController {
         }
 
         log.debug("Сформирован ответ {}", responseEntity);
+
+        Map<String, Object> monitoringParameters = new LinkedHashMap<>();
+        monitoringParameters.put("ID ", userCreate.getId());
+        monitoringParameters.put("Email ", userCreate.getEmail());
+        monitoringParameters.put("Имя ", userCreate.getUsername());
+        monitoringProducerService.send(
+                monitoringProducerService.constructReportDto(
+                        SecurityContextHolder.getContext().getAuthentication().getPrincipal(), ControlDto.MANUAL,
+                        this.getClass(), monitoringParameters));
+
         return responseEntity;
     }
 
@@ -162,6 +173,16 @@ public class UserController {
         }
 
         log.debug("Сформирован ответ {}", responseEntity);
+
+        Map<String, Object> monitoringParameters = new LinkedHashMap<>();
+        monitoringParameters.put("ID ", user.getId());
+        monitoringParameters.put("Email ", user.getEmail());
+        monitoringParameters.put("Имя ", user.getUsername());
+        monitoringProducerService.send(
+                monitoringProducerService.constructReportDto(
+                        SecurityContextHolder.getContext().getAuthentication().getPrincipal(), ControlDto.MANUAL,
+                        this.getClass(), monitoringParameters));
+
         return responseEntity;
     }
 
@@ -178,6 +199,12 @@ public class UserController {
             throw new RuntimeException("Method deleteUser is disabled by parameter deleteUserEnabled");
         }
 
+        User user = userService.getOneUser(id).orElse(null);
+        if (user == null) {
+            log.error("Пользователя с id = {} не существует", id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+
         ResponseEntity<User> responseEntity;
         try {
             userService.deleteUser(id);
@@ -186,6 +213,15 @@ public class UserController {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
         log.debug("Сформирован ответ {}", responseEntity);
+
+        Map<String, Object> monitoringParameters = new HashMap<>();
+        monitoringParameters.put("ID ", user.getId());
+        monitoringParameters.put("Email ", user.getEmail());
+        monitoringParameters.put("Имя ", user.getUsername());
+        monitoringProducerService.send(
+                monitoringProducerService.constructReportDto(
+                        SecurityContextHolder.getContext().getAuthentication().getPrincipal(), ControlDto.MANUAL,
+                        this.getClass(), monitoringParameters));
 
         return responseEntity;
     }
